pyproject.toml
[tool.poetry]
name = "moneybox-tg-bot"
version = "0.1.0"
description = ""
authors = ["Constantine Nazarov <m-end-eleev@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^2.0.25"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
psycopg2 = "^2.9.9"
pydantic = "^1.10.13"
python-dotenv = "^1.0.0"
telebot = "^0.0.5"
moneybox = "^0.0.35"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
